Hiding the gif window

+++ EclipseFolder/src/iec/MarioGIF.java
@@ -24,7 +24,24 @@ import javax.swing.border.EmptyBorder;
 public class MarioGIF {
 	
 	static int chosenGif;
-	public static int iecGeneration = 1; 
+	public static int iecGeneration = 1;
+	//Frame parameters
+	static JFrame frame;
+	
+	public static void reset(){
+		
+		//Create new frame
+		frame = new JFrame(); 
+		
+		//Delete images
+		deleteGifs();
+		
+		//reset chosen chromosome number
+		MarioGIF.setChosenGif(-1);
+		
+		//Show window
+		MarioGIF.setVisibility(true);
+	}
 	
 	public static void runIEC(){
 		
@@ -34,8 +51,6 @@ public class MarioGIF {
 				
 				chosenGif = -1;
 				
-				//Frame parameters
-				JFrame frame = new JFrame(); 
 				//JLabel headerLabel = new JLabel("headerLabel",JLabel.CENTER );
 				JPanel contentPane = new JPanel();
 				
@@ -86,12 +101,28 @@ public class MarioGIF {
 	}
 	
 	public static void deleteGifs(){
+		
+		//Create path to directory
 		final File dir = new File("./db/gifs/");
-	    ArrayList<BufferedImage> images = new ArrayList<BufferedImage>(); 
-	    File[] imgFiles = dir.listFiles();
-	    for(int i = 0; i < 9; i++)
-	    	imgFiles[ i ].delete(); 
+	    
+		//Create arraylist
+		ArrayList<BufferedImage> images = new ArrayList<BufferedImage>(); 
+	    
+		//Get files from directory
+		File[] imgFiles = dir.listFiles();
+	    
+	    //Check if any files in directory
+	    if( imgFiles.length > 0 )
+	    	//Go through all of them
+	    	for(int i = 0; i < imgFiles.length; i++)
+	    		//Delete file
+	    		imgFiles[ i ].delete(); 
 	}
+	
+	public static void setVisibility(boolean visible){
+		frame.setVisible(visible);
+	}
+	
 	
 	public static void setChosenGif(int buttonNumber){
 		chosenGif = buttonNumber;
